define the reco quantities in an additional file
